+++ myokit/tests/test_datalog.py
@@ -693,6 +693,16 @@ class DataLogTest(unittest.TestCase):
             list(d.keys()),
             ['time', 'x-1', 'x-3', 'ys-2-1', 'x-4', 'ys-2-2', 'x-2'])
 
+    def test_load_csv_ufeff(self):
+        # Test ignoring byte order marker (BOM) at start of file
+
+        with TemporaryDirectory() as td:
+            path = td.path('test.csv')
+            with open(path, 'w', encoding='utf-8-sig') as f:
+                f.write('time,x\n0,2\n1,3\n')
+            d = myokit.DataLog.load_csv(path)
+            self.assertEqual(list(d.keys()), ['time', 'x'])
+
     def test_load_csv_errors(self):
         # Test for errors during csv loading.